move quote role handling
